+++ src/context/builder.rs
@@ -64,10 +64,9 @@ impl LongPoolingServiceContextBuilder {
     /// use axum_cometd::LongPoolingServiceContextBuilder;
     ///
     /// let context = LongPoolingServiceContextBuilder::new().build();
-    /// # let app = axum_cometd::RouterBuilder::new().build::<()>(&context);
     /// ```
     #[inline(always)]
-    pub fn build<Msg>(self) -> Arc<LongPoolingServiceContext<Msg>> {
+    pub fn build(self) -> Arc<LongPoolingServiceContext> {
         let Self {
             subscriptions_storage_capacity,
             client_ids_storage_capacity,